requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.9.1
python-dotenv==1.0.1
requests==2.32.3
supabase==2.3.4
python-telegram-bot==22.0
google-generativeai==0.3.2
tenacity==8.2.3
